pyproject.toml
[project]
name = "wethrereport"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "ryouhama", email = "hary0730cap@icloud.com" }
]
dependencies = [
    "langchain>=0.0.27",
    "python-dotenv>=1.0.0",
    "pydantic>=1.10.9",
    "openai>=0.27.8",
    "pytest>=7.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true
